+++ src/cst/List.zig
@@ -16,9 +16,9 @@ pub fn init(self: *List, alloc: std.mem.Allocator) !void {
     self.items = try std.ArrayList(item.Item).initCapacity(alloc, 4);
 }
 
-pub fn getNum(self: List, index: usize) !?f32 {
+pub fn getNum(self: List, index: usize) !?f64 {
     if (self.items.items[index] == .number) {
-        return try std.fmt.parseFloat(f32, self.items.items[index].number.t_value.value);
+        return try std.fmt.parseFloat(f64, self.items.items[index].number.t_value.value);
     }
     return null;
 }
@@ -28,8 +28,8 @@ pub fn getVec2(self: List, index: usize) !?Vector2 {
         const i = self.items.items[index].list.items.items;
         if (i.len == 2 and i[0] == .number and i[1] == .number) {
             return .{
-                .x = try std.fmt.parseFloat(f32, i[0].number.t_value.value),
-                .y = try std.fmt.parseFloat(f32, i[1].number.t_value.value),
+                .x = try std.fmt.parseFloat(f64, i[0].number.t_value.value),
+                .y = try std.fmt.parseFloat(f64, i[1].number.t_value.value),
             };
         }
     }
@@ -41,9 +41,9 @@ pub fn getVec3(self: List, index: usize) !?Vector3 {
         const i = self.items.items[index].list.items.items;
         if (i.len == 3 and i[0] == .number and i[1] == .number and i[2] == .number) {
             return .{
-                .x = try std.fmt.parseFloat(f32, i[0].number.t_value.value),
-                .y = try std.fmt.parseFloat(f32, i[1].number.t_value.value),
-                .z = try std.fmt.parseFloat(f32, i[2].number.t_value.value),
+                .x = try std.fmt.parseFloat(f64, i[0].number.t_value.value),
+                .y = try std.fmt.parseFloat(f64, i[1].number.t_value.value),
+                .z = try std.fmt.parseFloat(f64, i[2].number.t_value.value),
             };
         }
     }
